fixed up some of the css
